innopoints_api/backend/requirements.txt
# Django
Django==2.1
uWSGI==2.0.17
psycopg2-binary==2.7.4
Pillow==5.2

# DRF
djangorestframework

# other
pytz
pycbrf==0.2.0

# Faker
factory_boy==2.11.1
